Move NL LinkbaseType to shared location
