+++ arelle/LinkbaseType.py
